Move verify to it's own file
